+++ Core/SoarKernel/src/xml.cpp
@@ -132,7 +132,7 @@ void xml_move_current_to_last_child( agent* pAgent )
 void xml_att_val( agent* pAgent, char const* pAttribute, uint64_t value )
 {
 	char buf[51];
-	SNPRINTF(buf, 50, "%llu", value);
+	SNPRINTF(buf, 50, "%llu", static_cast<long long unsigned>(value));
 
 	soarxml::XMLTrace* pXML = static_cast< soarxml::XMLTrace* >( pAgent->xml_destination );
 	pXML->AddAttribute( pAttribute, buf ) ;
@@ -150,7 +150,7 @@ void xml_att_val( agent* pAgent, char const* pAttribute, int value )
 void xml_att_val( agent* pAgent, char const* pAttribute, int64_t value )
 {
 	char buf[51];
-	SNPRINTF(buf, 50, "%lld", value);
+	SNPRINTF(buf, 50, "%lld", static_cast<long long>(value));
 
 	soarxml::XMLTrace* pXML = static_cast< soarxml::XMLTrace* >( pAgent->xml_destination );
 	pXML->AddAttribute( pAttribute, buf ) ;
